requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pydantic==2.10.5
pydantic_core==2.27.2
python-decouple==3.8
python-telegram-bot[job-queue]==21.10
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
